COAP-40: First example implementation of ACE client
Assumes Authorization Server (AS) is listening on a loopback interface.
Client requests access to SCOPE resource without knowing where this
resource is located. The IPv6 address of the RS is communicated to the
client in the audience parameter of the response. Upon reception of the
response, client installs a new OSCORE context and uses it to access the
SCOPE resource on the RS.